ice_cream_agent/requirements.txt
google-adk>=1.0.0
google-genai>=1.16.1